+++ src/components/app/app.scss
@@ -1,6 +1,6 @@
 .app {
   display: flex;
-  height: 101vh;
+  height: 100vh;
 
   &__container {
     width: 100%;
