+++ src/index.scss
@@ -24,6 +24,6 @@
 @import "./components/toolbar/toolbar";
 @import "./components/add-transaction-dialog/add-transaction-dialog";
 @import "./components/add-account-dialog/add-account-dialog";
-@import "components/buttonMore/button-more";
+@import "components/button-more/button-more";
 @import "./components/table-transactions/table-transactions";
 @import "./components/chart/chart";
